Extract the "memory" and "ttl" inline modules into their own files, as well as other helper methods in the "output" module
